+++ packages/sphinx-test-reports/docs/changelog.rst
@@ -50,7 +50,7 @@ Changelog
 
 0.3.0
 -----
-* Improvement: Using `sphinx-needs <https://sphinxcontrib-needs.readthedocs.io/en/latest/>`_ for data representation
+* Improvement: Using `sphinx-needs <https://sphinx-needs.readthedocs.io/en/latest/>`_ for data representation
   and filtering.
 * Improvement: New directives :ref:`test-file`, :ref:`test-suite` and :ref:`test-case`.
 * Improvement: New possibilities to :ref:`filter test data <filter>`.